18Ireland privates: Board of Works & Drumglass Colliery Railway
Fix Irish mail, P2 and P5 for narrow gaugeness
